Reorganize pages and routes
