+++ src/pages/Home/index.jsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { useNavigate } from 'react-router-dom'
-import { useNotification } from '../context/NotificationContext.jsx'
+import { useNotification } from '../../context/NotificationContext.jsx'
 
 export default function Home() {
   const navigate = useNavigate()
